preparations for gc; some fixes, avl tree map

+++ src/execution.h
@@ -2,7 +2,9 @@
 #ifndef ketl_execution_h
 #define ketl_execution_h
 
+#include "ketl/utils.h"
 
-int ketl_execute(void* opcodes);
+
+int64_t ketl_execute(void* opcodes);
 
 #endif // ketl_execution_h
